+++ sprints/admin.py
@@ -1,8 +1,10 @@
 from django.conf import settings
 from django.contrib import admin, messages
 from django.utils.html import format_html
+from django.utils import timezone
 
-from .models import Sprint
+from .models import Sprint, SprintMember
+from .services import SprintCloneService
 from .tasks import analyze_sprint
 
 
@@ -12,22 +14,28 @@ def analyze_current_sprint(modeladmin, request, queryset):
 analyze_current_sprint.short_description = 'Analyze current sprint'
 
 
-class SprintMemberInline(admin.TabularInline):
-    model = Sprint.members.through
+def generate_next_sprints(modeladmin, request, queryset):
+    for sprint in queryset:
+        SprintCloneService.clone_sprint(sprint)
+
+generate_next_sprints.short_description = 'Generate sprints from selected'
+
+
+class SprintMemberAdmin(admin.ModelAdmin):
+    pass
 
 
 class SprintAdmin(admin.ModelAdmin):
-    inlines = [SprintMemberInline]
-    actions = [analyze_current_sprint]
+    actions = [analyze_current_sprint, generate_next_sprints]
     list_display = (
         'project',
-        'number_team_members',
+        'number_team_members', 'member_names',
         'started_at', 'number_weeks', 'finished_at',
         'stories_done',
         'percentage_stories_done_bar',
         'number_releases', 'last_release'
     )
-    list_filter = ('started_at', 'project')
+    list_filter = ('started_at', 'project', 'members')
 
     def percentage_stories_done_bar(self, obj):
         return format_html(
@@ -40,3 +48,4 @@ class SprintAdmin(admin.ModelAdmin):
 
 
 admin.site.register(Sprint, SprintAdmin)
+admin.site.register(SprintMember, SprintMemberAdmin)